requirements.txt
flask==2.0.1
qrcode[pil]==7.3.1
pillow==9.0.0